third_party/WebKit/LayoutTests/fast/canvas/offscreencanvas-text/text/2d.text.font.parse.invalid.worker-expected.txt

This is a testharness.js-based test.
FAIL Untitled assert_equals: ctx.font === '20px serif' (got 0px [string], expected 20px serif[string]) expected "20px serif" but got "0px "
Harness: the test ran to completion.
